Use a function() enclosure for Closure-compiled code.
This is necessary because the closure compiler assumes
it is the only one using the global namespace. For example,
it sets "r" to null and uses r instead of null. Another
instance of the compiler might use "r" for something else.
That makes it impossible to paste together two separate libraries,
unless we enclose them in their own scope.